--- a/mimium-lang/src/compiler/bytecodegen.rs
+++ b/mimium-lang/src/compiler/bytecodegen.rs
@@ -6,8 +6,7 @@ use crate::mir::{self, Mir, StateSize};
 use crate::runtime::vm::bytecode::{ConstPos, GlobalPos, Reg};
 use crate::runtime::vm::{self};
 use crate::types::{PType, Type, TypeSize};
-use crate::utils::error::ReportableError;
-use half::f16;
+use crate::utils::{error::ReportableError,half_float::HFloat};
 use vm::bytecode::Instruction as VmInstruction;
 
 #[derive(Debug, Default)]
@@ -325,9 +324,8 @@ impl ByteCodeGenerator {
                 ))
             }
             mir::Instruction::Float(n) => {
-                let half_f = f16::from_f64(*n);
                 let dst = self.get_destination(dst, 1);
-                if half_f.is_finite() {
+                if let Ok(half_f) = HFloat::try_from(*n){
                     Some(VmInstruction::MoveImmF(dst, half_f))
                 } else {
                     let pos = funcproto.add_new_constant(gen_raw_float(n));