--- a/mimium-lang/src/utils.rs
+++ b/mimium-lang/src/utils.rs
@@ -3,6 +3,7 @@ pub mod error;
 pub mod fileloader;
 pub mod metadata;
 pub mod miniprint;
+pub mod half_float;
 
 #[macro_export]
 macro_rules! format_vec {