closes #77

Added Immidiate float operation for VM instruction.

Benchmark result:

```
with MoveImmF

running 7 tests
test tests::parse::bench_many_symbols10 ... bench:   8,342,941.60 ns/iter (+/- 1,049,541.66)
test tests::parse::bench_many_symbols3  ... bench:   1,085,331.25 ns/iter (+/- 63,756.41)
test tests::parse::bench_many_symbols5  ... bench:   2,491,700.00 ns/iter (+/- 150,930.81)
test tests::runtime::bench_multiosc10   ... bench:       3,498.56 ns/iter (+/- 106.64)
test tests::runtime::bench_multiosc15   ... bench:       5,238.47 ns/iter (+/- 438.19)
test tests::runtime::bench_multiosc5    ... bench:       1,775.72 ns/iter (+/- 39.10)
test tests::runtime::bench_multiosc7    ... bench:       2,468.23 ns/iter (+/- 70.50)

without 

running 7 tests
test tests::parse::bench_many_symbols10 ... bench:   8,527,756.30 ns/iter (+/- 623,910.55)
test tests::parse::bench_many_symbols3  ... bench:   1,100,597.90 ns/iter (+/- 17,084.55)
test tests::parse::bench_many_symbols5  ... bench:   2,525,233.40 ns/iter (+/- 62,603.26)
test tests::runtime::bench_multiosc10   ... bench:       3,596.35 ns/iter (+/- 975.28)
test tests::runtime::bench_multiosc15   ... bench:       5,373.25 ns/iter (+/- 124.99)
test tests::runtime::bench_multiosc5    ... bench:       1,819.19 ns/iter (+/- 53.49)
test tests::runtime::bench_multiosc7    ... bench:       2,547.23 ns/iter (+/- 89.92)
```

Little bit better?